refact: refactor getAllTasks function and styles components
